MAINT: finish removing "import .stats" from mstats_basic.py
